move crit application to actor
